[Refactor] game-option-page.h をsystem/enums/ からgame-option/ へ移した
元々system/ に置いていた名残だったが、system/ 以下のクラスが使っていなかった
