+++ src/birth/auto-roller.cpp
@@ -2,6 +2,7 @@
 #include "birth/birth-stat.h"
 #include "birth/birth-util.h"
 #include "cmd-io/cmd-gameoption.h"
+#include "game-option/game-option-page.h"
 #include "io/input-key-acceptor.h"
 #include "locale/japanese.h"
 #include "main/sound-of-music.h"
@@ -10,7 +11,6 @@
 #include "player/player-personality.h"
 #include "player/player-sex.h"
 #include "player/player-status-table.h"
-#include "system/enums/game-option-page.h"
 #include "system/player-type-definition.h"
 #include "term/screen-processor.h"
 #include "term/term-color-types.h"
