+++ src/cmd-io/cmd-gameoption.cpp
@@ -6,6 +6,7 @@
 #include "core/show-file.h"
 #include "core/window-redrawer.h"
 #include "floor/geometry.h"
+#include "game-option/game-option-page.h"
 #include "game-option/game-play-options.h"
 #include "game-option/keymap-directory-getter.h"
 #include "game-option/option-flags.h"
@@ -14,7 +15,6 @@
 #include "io/input-key-acceptor.h"
 #include "io/write-diary.h"
 #include "main/sound-of-music.h"
-#include "system/enums/game-option-page.h"
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "term/gameterm.h"
