+++ src/game-option/game-option-page.h
