+++ src/mind/mind-elementalist.cpp
@@ -19,6 +19,7 @@
 #include "floor/floor-util.h"
 #include "floor/geometry.h"
 #include "game-option/disturbance-options.h"
+#include "game-option/game-option-page.h"
 #include "game-option/input-options.h"
 #include "game-option/text-display-options.h"
 #include "grid/grid.h"
@@ -53,7 +54,6 @@
 #include "spell-kind/spells-world.h"
 #include "status/bad-status-setter.h"
 #include "status/base-status.h"
-#include "system/enums/game-option-page.h"
 #include "system/enums/terrain/terrain-characteristics.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
